create useAddItem hook that add item tofirestore
